Morlana_backend/requirements.txt
dotenv
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.2
coverage==7.11.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.119.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.0.1
ipython==8.37.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.15.1
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websocket-client==1.9.0
triton
nvidia-cufile-cu12==1.13.1.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90